Add new functions

+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/FilmDbStorage.java
@@ -1,32 +1,34 @@
 package ru.yandex.practicum.filmorate.storage.film;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Primary;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.BatchPreparedStatementSetter;
 import org.springframework.jdbc.core.JdbcTemplate;
+import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Component;
-import ru.yandex.practicum.filmorate.exceptions.NotFoundException;
+import ru.yandex.practicum.filmorate.exceptions.DuplicateEventException;
+import ru.yandex.practicum.filmorate.exceptions.*;
 import ru.yandex.practicum.filmorate.mappers.FilmMapper;
+import ru.yandex.practicum.filmorate.model.Director;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
+import ru.yandex.practicum.filmorate.model.enums.SearchParam;
 
 import java.sql.Date;
 import java.sql.PreparedStatement;
 import java.sql.SQLException;
 import java.util.*;
 
+@Primary
 @Component("filmDbStorage")
 @Slf4j
+@RequiredArgsConstructor
 public class FilmDbStorage implements FilmStorage {
     private final JdbcTemplate jdbcTemplate;
-
-    @Autowired
-    public FilmDbStorage(JdbcTemplate jdbcTemplate) {
-        this.jdbcTemplate = jdbcTemplate;
-    }
 
     @Override
     public List<Film> getAll() {
@@ -74,10 +76,11 @@ public class FilmDbStorage implements FilmStorage {
             return stmt;
         }, keyHolder);
 
-        Optional<Integer> id = Optional.of(keyHolder.getKey().intValue());
-        film.setId(id.get());
+        film.setId(keyHolder.getKey().intValue());
 
         setGenres(film);
+
+        setDirectorsToFilm(film);
 
         return film;
     }
@@ -102,7 +105,31 @@ public class FilmDbStorage implements FilmStorage {
 
         setGenres(film);
 
-        return findById(film.getId());
+        setDirectorsToFilm(film);
+
+        return film;
+    }
+
+    private void setDirectorsToFilm(Film film) {
+        String sqlQuery = "DELETE FROM FILMS_DIRECTORS WHERE FILM_ID = ?";
+        jdbcTemplate.update(sqlQuery, film.getId());
+        if (!film.getDirectors().isEmpty()) {
+            List<Director> directors = new ArrayList<>(film.getDirectors());
+            sqlQuery = "MERGE INTO FILMS_DIRECTORS (FILM_ID, DIRECTOR_ID) VALUES (?, ?)";
+
+            jdbcTemplate.batchUpdate(sqlQuery, new BatchPreparedStatementSetter() {
+                @Override
+                public void setValues(PreparedStatement ps, int i) throws SQLException {
+                    ps.setInt(1, film.getId());
+                    ps.setInt(2, directors.get(i).getId());
+                }
+
+                @Override
+                public int getBatchSize() {
+                    return directors.size();
+                }
+            });
+        }
     }
 
     private void setGenres(Film film) {
@@ -128,14 +155,29 @@ public class FilmDbStorage implements FilmStorage {
     }
 
     @Override
-    public void deleteAll() {
-        String query = "DELETE FROM FILMS";
-        jdbcTemplate.update(query);
+    public void deleteById(Integer id) {
+        this.isFilmExists(id);
+        String query = "delete from films where id = ?";
+        jdbcTemplate.update(query, id);
     }
 
     @Override
     public void addLike(Integer filmId, Integer userId) {
         this.isFilmExists(filmId);
+
+        String checkQuery =
+                "SELECT film_id " +
+                        "FROM films_likes " +
+                        "WHERE user_id = ? " +
+                        "AND film_id = ?;";
+
+        if (!jdbcTemplate
+                .query(checkQuery, (rs, n) -> rs.getInt("film_id"), userId, filmId)
+                .isEmpty()) {
+            throw new DuplicateEventException(String.format("User with id %d already liked film with id %d",
+                    filmId, userId));
+        }
+
         String query = "merge into films_likes(film_id, user_id) " +
                 "values (?, ?)";
         jdbcTemplate.update(query, filmId, userId);
@@ -144,11 +186,27 @@ public class FilmDbStorage implements FilmStorage {
                 "set rate = rate + 1 \n " +
                 "where id = ?";
         jdbcTemplate.update(updateFilmRate, filmId);
+
+        log.info("like for film with id={} added", filmId);
     }
 
     @Override
     public void deleteLike(Integer filmId, Integer userId) {
         this.isFilmExists(filmId);
+
+        String checkQuery =
+                "SELECT film_id " +
+                        "FROM films_likes " +
+                        "WHERE user_id = ? " +
+                        "AND film_id = ?;";
+
+        if (jdbcTemplate
+                .query(checkQuery, (rs, n) -> rs.getInt("film_id"), userId, filmId)
+                .isEmpty()) {
+            throw new NotFoundException("film", String.format("like from user with id %d to film with id %d not found",
+                    userId, filmId));
+        }
+
         String query = "delete from films_likes where film_id = ? and user_id = ?";
         jdbcTemplate.update(query, filmId, userId);
 
@@ -156,45 +214,134 @@ public class FilmDbStorage implements FilmStorage {
                 "set rate = rate - 1 \n " +
                 "where id = ?";
         jdbcTemplate.update(updateFilmRate, filmId);
+
+        log.info("like for film with id={} deleted", filmId);
     }
 
     @Override
-    public List<Film> getFilmsTop(Integer count) {
-        String query = "SELECT f.ID, f.NAME, f.DESCRIPTION, f.RELEASE_DATE, f.DURATION, f.RATE, f.MPA_RATE_ID, mr.name as mpa_name \n " +
+    public List<Film> getFilmsTop(Integer count, Integer genreId, Integer year) {
+        String whereYear = year != -1 ? " AND EXTRACT(YEAR FROM f.RELEASE_DATE) = " + year : "";
+        String joinGenres = genreId != -1 ? " left join FILMS_GENRES FG on f.ID = FG.FILM_ID " : "";
+        String whereGenre = genreId != -1 ? " AND FG.GENRE_ID = " + genreId : "";
+        String query = "SELECT f.*, mr.name as mpa_name \n " +
                 "FROM FILMS f \n " +
-                "left join MPA_RATING MR on f.MPA_RATE_ID = MR.MPA_RATE_ID \n " +
-                "ORDER BY f.rate DESC, f.ID " +
+                "left join MPA_RATING MR on f.MPA_RATE_ID = MR.MPA_RATE_ID \n " + joinGenres +
+                "WHERE 1=1 " + whereGenre + whereYear +
+                " ORDER BY f.rate DESC, f.ID " +
                 "LIMIT ? ";
 
         List<Film> filmsSorted = jdbcTemplate.query(query, FilmMapper::mapToFilm, count);
-
         this.setAttributes(filmsSorted);
-
         return filmsSorted;
+    }
+
+    @Override
+    public List<Film> searchFilm(String filter, List<SearchParam> params) {
+        String sqlQuery;
+        List<Film> films;
+
+        if (params.size() == 1) {
+            switch (params.get(0)) {
+                case TITLE:
+                    sqlQuery = "SELECT f.*, mr.name as mpa_name " +
+                            "FROM films f " +
+                            "JOIN mpa_rating mr ON f.mpa_rate_id = mr.mpa_rate_id " +
+                            "WHERE LOWER(f.name) LIKE LOWER(?)" +
+                            "ORDER BY f.rate DESC";
+                    break;
+
+                case DIRECTOR:
+                    sqlQuery = "SELECT f.*, mr.name as mpa_name " +
+                            "FROM directors d " +
+                            "JOIN films_directors fd ON d.id = fd.director_id " +
+                            "JOIN films f ON fd.film_id = f.id " +
+                            "JOIN mpa_rating mr ON f.mpa_rate_id = mr.mpa_rate_id " +
+                            "WHERE LOWER(d.name) LIKE LOWER(?)" +
+                            "GROUP BY f.id " +
+                            "ORDER BY f.rate DESC";
+                    break;
+
+                default:
+                    throw new IllegalArgumentException("incorrect filter type");
+            }
+
+            films = jdbcTemplate.query(sqlQuery, FilmMapper::mapToFilm, "%" + filter + "%");
+
+        } else {
+            sqlQuery = "SELECT f.*, mr.name as mpa_name " +
+                    "FROM directors d " +
+                    "JOIN films_directors fd ON d.id = fd.director_id " +
+                    "RIGHT JOIN films f ON fd.film_id = f.id " +
+                    "JOIN mpa_rating mr ON f.mpa_rate_id = mr.mpa_rate_id " +
+                    "WHERE LOWER(d.name) LIKE LOWER(?) OR LOWER(f.name) LIKE LOWER(?) " +
+                    "GROUP BY f.id " +
+                    "ORDER BY rate DESC";
+
+            films = jdbcTemplate.query(sqlQuery, FilmMapper::mapToFilm, "%" + filter + "%", "%" + filter + "%");
+        }
+
+        this.setAttributes(films);
+
+        return films;
+    }
+
+    public List<Film> getCommonFilms(Integer userId, Integer friendsId) {
+        String sqlQuery = "SELECT f.*, mr.name as mpa_name " +
+                "FROM films_likes fl " +
+                "JOIN films f ON fl.film_id = f.id " +
+                "JOIN mpa_rating mr ON f.mpa_rate_id = mr.mpa_rate_id " +
+                "WHERE fl.user_id = ? OR fl.user_id = ?" +
+                "GROUP BY fl.film_id " +
+                "HAVING COUNT(fl.film_id) > 1 " +
+                "ORDER BY f.rate DESC;";
+
+        List<Film> films = jdbcTemplate.query(sqlQuery, FilmMapper::mapToFilm, userId, friendsId);
+        this.setAttributes(films);
+
+        return films;
     }
 
     private void setAttributes(List<Film> films) {
         Map<Integer, Film> filmMap = new HashMap<>();
         films.forEach(film -> filmMap.put(film.getId(), film));
 
-        String genres = "SELECT fg.film_id, g.* " +
-                "FROM films_genres fg " +
-                "LEFT JOIN genres g on fg.genre_id = g.genre_id " +
-                "ORDER BY fg.FILM_ID, fg.GENRE_ID";
+        if (!filmMap.isEmpty()) {
 
-        jdbcTemplate.query(genres, rs -> {
-            Genre genre = new Genre();
-            genre.setId(rs.getInt("genre_id"));
-            genre.setName(rs.getString("name"));
-            filmMap.get(rs.getInt("film_id")).getGenres().add(genre);
-        });
+            String genres = "SELECT fg.film_id, g.* " +
+                    "FROM films_genres fg " +
+                    "LEFT JOIN genres g on fg.genre_id = g.genre_id " +
+                    "ORDER BY fg.FILM_ID, fg.GENRE_ID";
 
-        String likes = "SELECT * " +
-                "FROM films_likes";
+            jdbcTemplate.query(genres, rs -> {
+                Genre genre = new Genre();
+                genre.setId(rs.getInt("genre_id"));
+                genre.setName(rs.getString("name"));
+                Optional.ofNullable(filmMap.get(rs.getInt("film_id")))
+                        .ifPresent(f -> f.getGenres().add(genre));
+            });
 
-        jdbcTemplate.query(likes, rs -> {
-            filmMap.get(rs.getInt("film_id")).addLike(rs.getInt("user_id"));
-        });
+            String directorsSql = "SELECT FD.FILM_ID, D.* " +
+                    "FROM FILMS_DIRECTORS FD " +
+                    "LEFT JOIN DIRECTORS D ON FD.DIRECTOR_ID = D.ID " +
+                    "ORDER BY FD.FILM_ID, FD.DIRECTOR_ID";
+
+            jdbcTemplate.query(directorsSql, rs -> {
+                Director director = new Director();
+                director.setId(rs.getInt("id"));
+                director.setName(rs.getString("name"));
+                Optional.ofNullable(filmMap.get(rs.getInt("film_id")))
+                        .ifPresent(f -> f.getDirectors().add(director));
+            });
+
+            String likes = "SELECT * " +
+                    "FROM films_likes";
+
+            jdbcTemplate.query(likes, rs -> {
+                Integer userId = rs.getInt("user_id");
+                Optional.ofNullable(filmMap.get(rs.getInt("film_id")))
+                        .ifPresent(f -> f.addLike(userId));
+            });
+        }
     }
 
     private void setAttributes(Film film) {
@@ -209,6 +356,19 @@ public class FilmDbStorage implements FilmStorage {
             genre.setId(rs.getInt("genre_id"));
             genre.setName(rs.getString("name"));
             film.getGenres().add(genre);
+        }, film.getId());
+
+        String directorsSql = "SELECT FD.FILM_ID, D.* " +
+                "FROM FILMS_DIRECTORS FD " +
+                "LEFT JOIN DIRECTORS D ON FD.DIRECTOR_ID = D.ID " +
+                "WHERE FD.FILM_ID = ? " +
+                "ORDER BY FD.FILM_ID, FD.DIRECTOR_ID ";
+
+        jdbcTemplate.query(directorsSql, rs -> {
+            Director director = new Director();
+            director.setId(rs.getInt("id"));
+            director.setName(rs.getString("name"));
+            film.getDirectors().add(director);
         }, film.getId());
 
         String likes = "SELECT * " +
@@ -227,5 +387,66 @@ public class FilmDbStorage implements FilmStorage {
             throw new NotFoundException("id", String
                     .format("film with id %d does not exists", id));
         }
+    }
+
+
+    @Override
+    public List<Film> getFilmsByDirector(int directorId, String sortBy) {
+        List<Film> films;
+        switch (sortBy) {
+            case "year":
+                String sqlQueryYear = "SELECT F.*, MR.NAME as MPA_NAME\n" +
+                        "FROM FILMS F\n" +
+                        "LEFT JOIN MPA_RATING MR on F.MPA_RATE_ID = MR.MPA_RATE_ID\n" +
+                        "LEFT JOIN FILMS_DIRECTORS FD on F.ID = FD.FILM_ID\n" +
+                        "WHERE FD.DIRECTOR_ID = ?\n" +
+                        "ORDER BY F.RELEASE_DATE ASC, F.ID";
+                films = jdbcTemplate.query(sqlQueryYear, FilmMapper::mapToFilm, directorId);
+                break;
+            case "likes":
+                String sqlQueryLikes = "SELECT F.*, MR.NAME as MPA_NAME " +
+                        "FROM FILMS F " +
+                        "LEFT JOIN MPA_RATING MR on F.MPA_RATE_ID = MR.MPA_RATE_ID " +
+                        "LEFT JOIN FILMS_DIRECTORS FD on F.ID = FD.FILM_ID " +
+                        "WHERE FD.DIRECTOR_ID = ? " +
+                        "GROUP BY F.ID, MR.NAME " +
+                        "ORDER BY F.RATE DESC, F.ID";
+                films = jdbcTemplate.query(sqlQueryLikes, FilmMapper::mapToFilm, directorId);
+                break;
+            default:
+                throw new IllegalArgumentException("Invalid request parameter");
+        }
+        this.setAttributes(films);
+        return films;
+    }
+
+    @Override
+    public List<Integer> getRecommendations(Integer[] usersWithSimilarInterests, Integer idRecommendedUser,
+                                            Integer limit) {
+
+        String sql = "SELECT DISTINCT fl_1.film_id " +
+                "FROM films_likes AS fl_1 " +
+                "WHERE fl_1.user_id IN (?) " +
+                "EXCEPT " +
+                "SELECT fl_2.film_id " +
+                "FROM films_likes AS fl_2 " +
+                "WHERE fl_2.user_id = ? " +
+                "LIMIT ?";
+        return jdbcTemplate.queryForList(sql, Integer.class, usersWithSimilarInterests, idRecommendedUser, limit);
+
+    }
+
+    @Override
+    public List<Film> getFilmsByIdList(List<Integer> idList) {
+        String inClause = String.join(",", Collections.nCopies(idList.size(), "?"));
+        String query = String.format("SELECT f.*, mr.name as mpa_name " +
+                "FROM films f " +
+                "left join MPA_RATING MR on f.MPA_RATE_ID = MR.MPA_RATE_ID " +
+                "WHERE f.id in (%s)", inClause);
+
+        List<Film> films = jdbcTemplate.query(query, FilmMapper::mapToFilm, idList.toArray());
+        this.setAttributes(films);
+
+        return films;
     }
 }
